Async plugin: Refactor to use traits for ops (fixes #61)
This gets rid of all the parsing and emitting done by
operation inference by adding a bridge trait (OperationModifier)
which inherits from Apiv2Schema and is implemented for most
of the types. Finally, Apiv2Operation is implemented for functions
whose components implement this bridge trait. This also solves
the Factory vs AsyncFactory problem.