fix: handle parameters on the format r#type (#429)
A new version of my previous non working PR, this seems to actually do what I intended.

When you have a struct with a member using a reserved word as a name, it will be specified as r#type. For Query structs, this is probably quite common. Previously they got hte name "r#type" in the openapi spec, so I just added a strip_prefix r# to handle this case.

Co-authored-by: Mattias Eriksson <snaggen@mayam.com>