Set refresh_url to an empty string
The previous value of refresh_url (None) translated into 'null' in the
spec, which is not accepted as a proper value, as refresh_url is
supposed to have a string value.

To see an example of how this goes wrong, you can create a spec with the
old code (i.e. with refresh_url: None) and paste it into
editor.swagger.io. You'll get an error. Now generate a spec using this
new code (i.e. with refresh_url: "") and the error goes away.