build: update openapiv3 to v2.0
There's a breaking change on the crate which does not expose the fields as pub
and so we need to use our own patch for now.
In fact, we can't merge this upstream (not maintained) so instead we move to
a fork of our own.
Since this is not part of our public api this should be fairly self contained.

Signed-off-by: Tiago Castro <tiagolobocastro@gmail.com>